update counting rekap nilai

+++ database/migrations/2023_07_08_070650_create_pengajuan_sidangs_table.php
@@ -24,7 +24,7 @@ return new class extends Migration
             $table->string('dosen_penguji2')->nullable();
             $table->string('dosen_penguji3')->nullable();
             $table->string('status')->nullable();
-            $table->integer('nilai_pembimbing')->nullable();
+            $table->float('nilai_pembimbing')->nullable();
             $table->timestamps();
         });
     }
